Added background image and some styling to the front end

+++ api/game/reset.php
@@ -16,19 +16,33 @@
   // Instantiate army and game object
   $army = new Army($db);
   $game = new Game($db);
-
+  
   // Get IDs
   $army->game_id = isset($_GET['id']) ? $_GET['id'] : die();
   $game->id = isset($_GET['id']) ? $_GET['id'] : die();
 
-  // Reset game
-  if($game->reset() && $army->delete()) {
-    echo json_encode(
-      array('message' => 'Game successfuly reseted! Check game log!')
-    );
+  // Get games
+  $result = $game->get_all();
+  
+  // Get row count
+  $num = $result->rowCount();
+
+  // Check if teheres any games
+  if($num > 0) {
+
+    // Reset game
+    if($game->reset() && $army->delete()) {
+      echo json_encode(
+        array('message' => 'Game successfuly reseted! Check game log!')
+      );
+    } else {
+      echo json_encode(
+        array('message' => 'Something went wrong! Game not reseted!')
+      );
+    }
   } else {
     echo json_encode(
-      array('message' => 'Something went wrong! Game not reseted!')
+      array('message' => 'No games found!')
     );
   }
